Move DuckDb parquet reader to a subfolder
This commit moves the DuckDb parquet reader to a new folder
"duckdb_reader" under “dwio/parquet", to prepare for the upcoming
native parquet reader.
