+++ velox/dwio/CMakeLists.txt
@@ -30,6 +30,7 @@ set(VELOX_LINK_LIBS
 
 add_subdirectory(common)
 add_subdirectory(dwrf)
+
 if(VELOX_ENABLE_PARQUET)
   add_subdirectory(parquet)
 endif()
