+++ velox/dwio/parquet/duckdb_reader/CMakeLists.txt
@@ -12,7 +12,10 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-add_library(velox_dwio_parquet_reader_duckdb Allocator.cpp)
+add_subdirectory(duckdb)
 
-target_link_libraries(velox_dwio_parquet_reader_duckdb velox_dwio_common duckdb
-                      ${FMT})
+add_library(velox_dwio_duckdb_parquet_reader ParquetReader.cpp)
+
+target_link_libraries(
+  velox_dwio_duckdb_parquet_reader velox_dwio_duckdb_parquet_reader_duckdb
+  velox_dwio_common velox_duckdb_conversion duckdb ${FMT})
