+++ velox/dwio/parquet/duckdb_reader/ParquetReader.h
@@ -19,13 +19,13 @@
 #include "velox/common/base/Macros.h"
 #include "velox/dwio/common/Reader.h"
 #include "velox/dwio/common/ReaderFactory.h"
-#include "velox/dwio/parquet/reader/duckdb/Allocator.h"
-#include "velox/dwio/parquet/reader/duckdb/InputStreamFileSystem.h"
+#include "velox/dwio/parquet/duckdb_reader/duckdb/Allocator.h"
+#include "velox/dwio/parquet/duckdb_reader/duckdb/InputStreamFileSystem.h"
 VELOX_SUPPRESS_DEPRECATION_WARNING
 #include "velox/external/duckdb/parquet-amalgamation.hpp"
 VELOX_UNSUPPRESS_DEPRECATION_WARNING
 
-namespace facebook::velox::parquet {
+namespace facebook::velox::parquet::duckdb_reader {
 
 class ParquetRowReader : public dwio::common::RowReader {
  public:
@@ -95,8 +95,4 @@ class ParquetReaderFactory : public dwio::common::ReaderFactory {
   }
 };
 
-void registerParquetReaderFactory();
-
-void unregisterParquetReaderFactory();
-
-} // namespace facebook::velox::parquet
+} // namespace facebook::velox::parquet::duckdb_reader
